cheri: fs: smb: Fix uintptr_t usage in fs/smb/
Use uintptr_t for values that can contain a pointer in fs/smb/
and cast as necessary.

+++ fs/smb/client/cifsproto.h
@@ -516,7 +516,7 @@ static inline int cifs_get_num_sgs(const struct smb_rqst *rqst,
 {
 	unsigned int len, skip;
 	unsigned int nents = 0;
-	unsigned long addr;
+	uintptr_t addr;
 	size_t data_size;
 	int i, j;
