+++ fs/smb/client/smbdirect.c
@@ -3198,7 +3198,7 @@ static ssize_t smb_extract_kvec_to_rdma(struct iov_iter *iter,
 
 	for (i = 0; i < iter->nr_segs; i++) {
 		struct page *page;
-		unsigned long kaddr;
+		uintptr_t kaddr;
 		size_t off, len, seg;
 
 		len = kv[i].iov_len;
@@ -3207,7 +3207,7 @@ static ssize_t smb_extract_kvec_to_rdma(struct iov_iter *iter,
 			continue;
 		}
 
-		kaddr = (unsigned long)kv[i].iov_base + start;
+		kaddr = (uintptr_t)kv[i].iov_base + start;
 		off = kaddr & ~PAGE_MASK;
 		len = min_t(size_t, maxsize, len - start);
 		kaddr &= PAGE_MASK;
